feat: add validation for node type

+++ packages/oracle-manager/contracts/storage/NodeFactoryStorage.sol
@@ -20,7 +20,7 @@ contract NodeFactoryStorage {
     }
 
     struct NodeData {
-        int price;
+        int256 price;
         uint timestamp;
         uint volatilityScore;
         uint liquidityScore;
